moved most code in index.js to app.js
